Trimming the fat
Went through the entire code and strip it down to basics.  Took out a
lot of references to the bootstrap framework because I realized that
there wasn't anything special being utilized by it that I couldn't do in
HTML/CSS.  Deleted all the bootstrap CSS files and changed the CSS
references on all the pages to point to the one CSS file.